merge: 메인 하단·상단 작업 내용 통합
